Replacing previous script
Upgrading previous script as requested by @MiguelCasaresRobles
